+++ R/pb_download_url.R
@@ -1,25 +1,65 @@
 #' Get the download url of a given file
 #'
-#' Returns the URL download for a public file. This can be useful when writing
-#' scripts that may want to download the file directly without introducing any
-#' dependency on `piggyback` or authentication steps.
+#' Returns the URL download for a given file. This can be useful when using
+#' functions that are able to accept URLs.
+#'
+#' @param url_type choice: one of "browser" or "api" - default "browser" is a
+#' web-facing URL that is not subject to API ratelimits but does not work for
+#' private repositories. "api" URLs work for private repos, but require a GitHub
+#' token passed in an Authorization header (see examples)
 #' @inheritParams pb_download
 #' @return the URL to download a file
 #' @export
-#' @examples \dontrun{
+#' @examples \donttest{
+#' \dontshow{try(\{}
 #'
-#' pb_download_url("iris.tsv.xz",
-#'                 repo = "cboettig/piggyback-tests",
-#'                 tag = "v0.0.1")
+#' # returns browser url by default (and all files if none are specified)
+#' browser_url <- pb_download_url(
+#'   repo = "tanho63/piggyback-tests",
+#'   tag = "v0.0.2"
+#'   )
+#' print(browser_url)
+#' utils::read.csv(browser_url[[1]])
 #'
+#' # can return api url if desired
+#' api_url <- pb_download_url(
+#'   "mtcars.csv",
+#'   repo = "tanho63/piggyback-tests",
+#'   tag = "v0.0.2"
+#'   )
+#' print(api_url)
+#'
+#' # for public repositories, this will still work
+#' utils::read.csv(api_url)
+#'
+#' # for private repos, can use httr or curl to fetch and then pass into read function
+#' gh_pat <- Sys.getenv("GITHUB_PAT")
+#'
+#' if(!identical(gh_pat, "")){
+#'   resp <- httr::GET(api_url, httr::add_headers(Authorization = paste("Bearer", gh_pat)))
+#'   utils::read.csv(text = httr::content(resp, as = "text"))
+#' }
+#'
+#' # or use pb_read which bundles some of this for you
+#'
+#' \dontshow{\})}
 #' }
 pb_download_url <- function(file = NULL,
                             repo = guess_repo(),
                             tag = "latest",
+                            url_type = c("browser","api"),
                             .token = gh::gh_token()) {
+  url_type <- rlang::arg_match(url_type, values = c("browser","api"))
+
   df <- pb_info(repo, tag, .token)
 
-  if(is.null(file)) return(df$browser_download_url)
+  if(is.null(file)) {
+    switch(
+      url_type,
+      "browser" = return(df$browser_download_url),
+      "api" = return(df$api_download_url)
+    )
+  }
 
   if(any(!file %in% df$file_name)) {
 
@@ -32,5 +72,9 @@ pb_download_url <- function(file = NULL,
 
   if(length(file) == 0) return(cli::cli_abort("No download URLs to return."))
 
-  return(df[df$file_name %in% file,"browser_download_url"])
+  switch(
+    url_type,
+    "browser" = return(df$browser_download_url[df$file_name %in% file]),
+    "api" = return(df$api_download_url[df$file_name %in% file])
+  )
 }